+++ tests/test_api.py
@@ -2,7 +2,7 @@ import pytest
 
 from api_tabular.utils import external_url
 
-from .conftest import RESOURCE_ID, DATE, PG_RST_URL, TABLES_INDEX_PATTERN
+from .conftest import RESOURCE_ID, DATE, PGREST_ENDPOINT, TABLES_INDEX_PATTERN
 
 pytestmark = pytest.mark.asyncio
 
@@ -57,7 +57,7 @@ async def test_api_resource_profile_not_found(client, mock_get_resource_empty):
 async def test_api_resource_data(client, rmock):
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PG_RST_URL}/xxx?limit=20&order=__id.asc",
+        f"{PGREST_ENDPOINT}/xxx?limit=20&order=__id.asc",
         payload={"such": "data"},
         headers={"Content-Range": "0-10/10"},
     )
@@ -79,7 +79,7 @@ async def test_api_resource_data_with_args(client, rmock):
     args = "page=1"
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PG_RST_URL}/xxx?limit=20&order=__id.asc",
+        f"{PGREST_ENDPOINT}/xxx?limit=20&order=__id.asc",
         payload={"such": "data"},
         headers={"Content-Range": "0-10/10"},
     )
@@ -101,7 +101,7 @@ async def test_api_resource_data_with_args_case(client, rmock):
     args = "COLUM_NAME__EXACT=BIDULE&page=1"
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PG_RST_URL}/xxx?COLUM_NAME=eq.BIDULE&limit=20&order=__id.asc",
+        f"{PGREST_ENDPOINT}/xxx?COLUM_NAME=eq.BIDULE&limit=20&order=__id.asc",
         payload={"such": "data"},
         headers={"Content-Range": "0-10/10"},
     )
@@ -158,7 +158,7 @@ async def test_api_resource_data_not_found(client, mock_get_resource_empty):
 
 async def test_api_resource_data_table_error(client, rmock):
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
-    rmock.get(f"{PG_RST_URL}/xxx?limit=20&order=__id.asc", status=502, payload={"such": "error"})
+    rmock.get(f"{PGREST_ENDPOINT}/xxx?limit=20&order=__id.asc", status=502, payload={"such": "error"})
     res = await client.get(f"/api/resources/{RESOURCE_ID}/data/")
     assert res.status == 502
     assert await res.json() == {
@@ -171,7 +171,7 @@ async def test_api_resource_data_table_error(client, rmock):
 async def test_api_percent_encoding_arabic(client, rmock):
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PG_RST_URL}/xxx?%D9%85%D9%88%D8%A7%D8%B1%D8%AF=eq.%D9%85%D9%88%D8%A7%D8%B1%D8%AF&limit=20&order=__id.asc",
+        f"{PGREST_ENDPOINT}/xxx?%D9%85%D9%88%D8%A7%D8%B1%D8%AF=eq.%D9%85%D9%88%D8%A7%D8%B1%D8%AF&limit=20&order=__id.asc",  # noqa
         status=200,
         payload={"such": "data"},
         headers={"Content-Range": "0-10/10"},
@@ -195,7 +195,7 @@ async def test_api_percent_encoding_arabic(client, rmock):
 async def test_api_with_unsupported_args(client, rmock):
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PG_RST_URL}/xxx?limit=20&order=__id.asc",
+        f"{PGREST_ENDPOINT}/xxx?limit=20&order=__id.asc",
         status=200,
         payload={"such": "data"},
         headers={"Content-Range": "0-10/10"},
@@ -217,7 +217,7 @@ async def test_api_with_unsupported_args(client, rmock):
 async def test_api_pagination(client, rmock):
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PG_RST_URL}/xxx?limit=1&order=__id.asc",
+        f"{PGREST_ENDPOINT}/xxx?limit=1&order=__id.asc",
         status=200,
         payload=[{"such": "data"}],
         headers={"Content-Range": "0-2/2"},
@@ -238,7 +238,7 @@ async def test_api_pagination(client, rmock):
 
     rmock.get(TABLES_INDEX_PATTERN, payload=[{"__id": 1, "id": "test-id", "parsing_table": "xxx"}])
     rmock.get(
-        f"{PG_RST_URL}/xxx?limit=1&offset=1&order=__id.asc",
+        f"{PGREST_ENDPOINT}/xxx?limit=1&offset=1&order=__id.asc",
         status=200,
         payload=[{"such": "data"}],
         headers={"Content-Range": "0-2/2"},